components/core/requirements.txt
click==6.7
Flask==0.10.1
Flask-Cors==2.1.2
Flask-MySQLdb==0.2.0
Flask-SocketIO==2.6.1
gevent==1.1.0
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.7
python-dotenv==0.7.1
python-engineio==2.0.2
python-socketio==1.8.4
six==1.10.0
SQLAlchemy==1.1.9
websocket-client==0.37.0
Werkzeug==0.11.4
